Moved some permissions logic to the shared module
