+++ shared/src/commonMain/kotlin/org/centrexcursionistalcoi/app/security/ParametrizedPermission.kt
