+++ shared/src/commonTest/kotlin/org/centrexcursionistalcoi/app/security/TestParametrizedPermission.kt
@@ -10,7 +10,7 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission with single parameter`() {
         val permission = ParametrizedPermission("department.*.kick")
-        
+
         // Test with one parameter
         val result = permission("marketing")
         assertEquals("department.marketing.kick", result)
@@ -19,7 +19,7 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission with multiple parameters`() {
         val permission = ParametrizedPermission("resource.*.*.access")
-        
+
         // Test with multiple parameters
         val result = permission("documents", "confidential")
         assertEquals("resource.documents.confidential.access", result)
@@ -28,7 +28,7 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission with no parameters`() {
         val permission = ParametrizedPermission("user.view")
-        
+
         // Test with no parameters (should work with empty vararg)
         val result = permission()
         assertEquals("user.view", result)
@@ -37,7 +37,7 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission with wrong number of parameters`() {
         val permission = ParametrizedPermission("department.*.kick")
-        
+
         // Test with wrong number of parameters - should throw
         assertFailsWith<IllegalArgumentException> {
             permission("marketing", "extra")
@@ -47,7 +47,7 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission with too few parameters`() {
         val permission = ParametrizedPermission("resource.*.*.access")
-        
+
         // Test with too few parameters - should throw
         assertFailsWith<IllegalArgumentException> {
             permission("documents")
@@ -58,7 +58,7 @@ class TestParametrizedPermission {
     fun `test parametrized permission extension function`() {
         // Test the extension function
         val permission = "department.*.kick".parametrized()
-        
+
         val result = permission("finance")
         assertEquals("department.finance.kick", result)
     }
@@ -66,12 +66,12 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission with different parameter types`() {
         val permission = ParametrizedPermission("item.*.view")
-        
+
         // Test with different parameter types (should convert to string)
         val result1 = permission(123)
         val result2 = permission(true)
         val result3 = permission(42.5)
-        
+
         assertEquals("item.123.view", result1)
         assertEquals("item.true.view", result2)
         assertEquals("item.42.5.view", result3)
@@ -80,7 +80,7 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission with special characters in parameters`() {
         val permission = ParametrizedPermission("file.*.access")
-        
+
         // Test with special characters in parameters
         val result = permission("documents/confidential/file.txt")
         assertEquals("file.documents/confidential/file.txt.access", result)
@@ -89,12 +89,12 @@ class TestParametrizedPermission {
     @Test
     fun `test parametrized permission rejects blank parameters`() {
         val permission = ParametrizedPermission("group.*.access")
-        
+
         // Test that blank parameters are rejected
         assertFailsWith<IllegalArgumentException> {
             permission("")
         }
-        
+
         assertFailsWith<IllegalArgumentException> {
             permission("   ")
         }
@@ -105,12 +105,12 @@ class TestParametrizedPermission {
         val permission1 = ParametrizedPermission("a.*.b")
         val permission2 = ParametrizedPermission("a.*.b.*.c")
         val permission3 = ParametrizedPermission("a.b.c")
-        
+
         // These should work
         permission1("x")
         permission2("x", "y")
         permission3()
-        
+
         // These should fail
         assertFailsWith<IllegalArgumentException> { permission1() }
         assertFailsWith<IllegalArgumentException> { permission1("x", "y") }
@@ -124,10 +124,10 @@ class TestParametrizedPermission {
         // Test that the permissions defined in the Permissions object work correctly
         val kickPermission = Permissions.Department.KICK
         val manageRequestsPermission = Permissions.Department.MANAGE_REQUESTS
-        
+
         val kickResult = kickPermission("it")
         val manageResult = manageRequestsPermission("hr")
-        
+
         assertEquals("department.it.kick", kickResult)
         assertEquals("department.hr.manage_requests", manageResult)
     }
